Original change: https://android-review.googlesource.com/c/platform/tools/tradefederation/+/1428168

+++ src/com/android/tradefed/device/metric/JavaCodeCoverageCollector.java
@@ -36,6 +36,8 @@ import com.android.tradefed.testtype.coverage.CoverageOptions;
 import com.android.tradefed.util.AdbRootElevator;
 import com.android.tradefed.util.FileUtil;
 import com.android.tradefed.util.JavaCodeCoverageFlusher;
+import com.android.tradefed.util.ProcessInfo;
+import com.android.tradefed.util.PsParser;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Splitter;
@@ -45,6 +47,7 @@ import org.jacoco.core.tools.ExecFileLoader;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Map;
 import java.util.List;
 
@@ -162,7 +165,6 @@ public final class JavaCodeCoverageCollector extends BaseDeviceMetricCollector
                 devicePaths.addAll(Splitter.on('\n').omitEmptyStrings().split(fileList));
 
                 collectAndLogCoverageMeasurements(device, devicePaths.build());
-
             } catch (DeviceNotAvailableException | IOException e) {
                 throw new RuntimeException(e);
             }
@@ -177,10 +179,24 @@ public final class JavaCodeCoverageCollector extends BaseDeviceMetricCollector
 
     private void collectAndLogCoverageMeasurements(ITestDevice device, List<String> devicePaths)
             throws IOException, DeviceNotAvailableException {
+        List<Integer> activePids = getRunningProcessIds(device);
 
         for (String devicePath : devicePaths) {
             File coverageFile = device.pullFile(devicePath);
-            device.deleteFile(devicePath);
+
+            if (devicePath.endsWith(".mm.ec")) {
+                // Check if the process was still running. The file will have the format
+                // /data/misc/trace/jacoco-XXXXX.mm.ec where XXXXX is the process id.
+                int start = devicePath.indexOf('-') + 1;
+                int end = devicePath.indexOf('.');
+                int pid = Integer.parseInt(devicePath.substring(start, end));
+                if (!activePids.contains(pid)) {
+                    device.deleteFile(devicePath);
+                }
+            } else {
+                device.deleteFile(devicePath);
+            }
+
             verifyNotNull(
                     coverageFile, "Failed to pull the Java code coverage file from %s", devicePath);
 
@@ -201,6 +217,17 @@ public final class JavaCodeCoverageCollector extends BaseDeviceMetricCollector
                 FileUtil.deleteFile(coverageFile);
             }
         }
+    }
+
+    private List<Integer> getRunningProcessIds(ITestDevice device)
+            throws DeviceNotAvailableException {
+        List<ProcessInfo> processes = PsParser.getProcesses(device.executeShellCommand("ps -e"));
+        List<Integer> pids = new ArrayList<>();
+
+        for (ProcessInfo process : processes) {
+            pids.add(process.getPid());
+        }
+        return pids;
     }
 
     private FailureDescription createCodeCoverageFailure(String message) {
